[Refactor] Handle pathological strings better
Improve input string that are pathological by virtue
of containing lots of whitespace but not at the end
or start of the string. Such strings can lead to
quadratic runtime of the `rightWhitespace` regular
expression, unnecessarily consuming computational
resources.

This patches addresses the problem by ensuring that
the string does end with a whitespace character
before removing whitespace character from the end of
the string. This ensures that pathological inputs
cannot cause quadratic runtime because such inputs
won't reach the `rightWhitespace` expression.

An alternative solution would be to have a negative
lookbehind in the `rightWhitespace` expression. But,
this regexp feature was not yet available in versions
of JavaScript for which this polyfill is meant, hence
it is not viable.